use single component for all conversion error modals
