--- src/components/ConversionFileErrorModal/style.css
@@ -1,4 +1,0 @@
-.error-modal .warning-text {
-    color: var(--cancel-icon-color);
-    padding-bottom: 0.5em;
-}
