--- src/components/ConversionServerErrorModal/style.css
@@ -1,4 +1,0 @@
-.server-error-modal .warning-text {
-    color: var(--cancel-icon-color);
-    padding-bottom: 0.5em;
-}
